@@ -39,16 +39,7 @@ internal static class HttpClientHandlerHelper
             throw new InvalidOperationException("Underlying handler for HttpClientHandler is not SocketsHttpHandler");
         }
 
-        underlyingHandler.EnableMultipleHttp2Connections = true;
-
-        if (pooledConnectionLifetime == null && pooledConnectionIdleTimeout == null)
-            return;
-
-        if (pooledConnectionLifetime.HasValue)
-            underlyingHandler.PooledConnectionLifetime = pooledConnectionLifetime.Value;
-
-        if (pooledConnectionIdleTimeout.HasValue)
-            underlyingHandler.PooledConnectionIdleTimeout = pooledConnectionIdleTimeout.Value;
+        Configure(underlyingHandler, pooledConnectionLifetime, pooledConnectionIdleTimeout);
 #else
         if (pooledConnectionLifetime == null && pooledConnectionIdleTimeout == null)
             return;
@@ -56,4 +47,23 @@ internal static class HttpClientHandlerHelper
         throw new InvalidOperationException("Cannot set 'PooledConnectionLifetime' and 'PooledConnectionIdleTimeout' in this framework. Should not happen!");
 #endif
     }
+
+#if NETCOREAPP3_1_OR_GREATER
+    public static void Configure(SocketsHttpHandler handler, TimeSpan? pooledConnectionLifetime, TimeSpan? pooledConnectionIdleTimeout)
+    {
+        if (handler == null)
+            throw new ArgumentNullException(nameof(handler));
+
+        handler.EnableMultipleHttp2Connections = true;
+
+        if (pooledConnectionLifetime == null && pooledConnectionIdleTimeout == null)
+            return;
+
+        if (pooledConnectionLifetime.HasValue)
+            handler.PooledConnectionLifetime = pooledConnectionLifetime.Value;
+
+        if (pooledConnectionIdleTimeout.HasValue)
+            handler.PooledConnectionIdleTimeout = pooledConnectionIdleTimeout.Value;
+    }
+#endif
 }
