@@ -17,6 +17,7 @@ using FluentFTP;
 using FluentFTP.Client.BaseClient;
 using JetBrains.Annotations;
 using Raven.Client.Documents.Operations.Backups;
+using Raven.Client.Util;
 
 namespace Raven.Server.Documents.PeriodicBackup
 {
@@ -165,7 +166,7 @@ namespace Raven.Server.Documents.PeriodicBackup
 
                 try
                 {
-                    var x509Certificate = new X509Certificate2(byteArray);
+                    var x509Certificate = CertificateHelper.CreateCertificate(byteArray);
                     client.Config.EncryptionMode = FtpEncryptionMode.Explicit;
                     client.Config.ClientCertificates.Add(x509Certificate);
                     client.ValidateCertificate += OnValidateCertificate;
