@@ -3285,7 +3285,7 @@ namespace Raven.Server
                 {
                     try
                     {
-                        certificate = new X509Certificate2(buffer[0..read]);
+                        certificate = CertificateHelper.CreateCertificate(buffer[0..read]);
                     }
                     catch (Exception e)
                     {
@@ -3298,7 +3298,7 @@ namespace Raven.Server
                 {
                     try
                     {
-                        certificate = new X509Certificate2(issuer);
+                        certificate = CertificateHelper.CreateCertificate(issuer);
                     }
                     catch (Exception e)
                     {
