@@ -51,7 +51,7 @@ namespace SlowTests.Authentication
 
         public X509Certificate2 CreateAndPutExpiredClientCertificate(string serverCertPath, Dictionary<string, DatabaseAccess> permissions, SecurityClearance clearance = SecurityClearance.ValidUser)
         {
-            var serverCertificate = new X509Certificate2(serverCertPath, (string)null, X509KeyStorageFlags.MachineKeySet);
+            var serverCertificate = CertificateHelper.CreateCertificate(serverCertPath, (string)null, X509KeyStorageFlags.UserKeySet);
             var serverCertificateHolder = new SecretProtection(
                 new SecurityConfiguration()).LoadCertificateFromPath(
                 serverCertPath,
