@@ -8,6 +8,7 @@ using McMaster.Extensions.CommandLineUtils;
 using Newtonsoft.Json;
 using Raven.Client.Documents;
 using Raven.Client.ServerWide.Operations.Certificates;
+using Raven.Client.Util;
 using Raven.Server.Commercial;
 using Sparrow.Json;
 using Sparrow.Platform;
@@ -336,8 +337,8 @@ namespace rvn
                     }
 
                     
-                    X509Certificate2 clientCertificate = new(clientCertificatePathArg.Value);
-                    X509Certificate2 serverCertificate = new(serverCertificatePathArg.Value);
+                    X509Certificate2 clientCertificate = CertificateHelper.CreateCertificate(clientCertificatePathArg.Value);
+                    X509Certificate2 serverCertificate = CertificateHelper.CreateCertificate(serverCertificatePathArg.Value);
                     var name = Path.GetFileNameWithoutExtension(clientCertificatePathArg.Value);
                     try
                     {
